+++ app/resources/views/pages/dominion/retire.blade.php
@@ -99,7 +99,6 @@
                                         <th>Name</th>
                                         <th>Level Bonus</th>
                                         <th>Requirement</th>
-                                        <th>Starting XP</th>
                                         <th>Upgrades</th>
                                     </tr>
                                 </thead>
@@ -111,11 +110,12 @@
                                         <td>{{ $class['name'] }}</td>
                                         <td>{{ ucwords(str_replace('_', ' ', $class['perk_type'])) }}</td>
                                         <td>{{ $heroHelper->getRequirementDisplay($class) }}</td>
-                                        <td>{{ $heroHelper->getStartingExperienceDisplay($class) }}</td>
                                         <td>
                                             @foreach ($heroHelper->getHeroUpgradesByName($class['perks'])->where('active', true) as $upgrade)
                                                 {{ $upgrade->name }}
-                                                @if ($upgrade->type !== 'directive')
+                                                @if ($upgrade->type == 'directive')
+                                                    (required)
+                                                @else
                                                     (optional)
                                                 @endif
                                                 <br/>
@@ -201,7 +201,7 @@
                     <p>Your hero gains 1 XP per acre gained from invasion, 1 XP per successful info operation (excluding bots), 4 XP per successful black operation, and 6 XP per successful war operation.</p>
                     <p>Your hero loses 1 XP per acre lost from invasion, however this loss cannot exceed the XP required to maintain its current level.</p>
                     <p>Basic hero classes can be retired and replaced with another class. Basic hero classes will start with XP equal to half that of its predecessor.</p>
-                    <p>Advanced hero classes cannot be retired, cannot be selected until the 10th day of the round, and unlock additional upgrades.</p>
+                    <p>Advanced hero classes cannot be retired, cannot be selected until the 5th day of the round, and unlock additional upgrades.</p>
                 </div>
             </div>
         </div>