+++ app/resources/views/pages/scribes/heroes.blade.php
@@ -15,7 +15,7 @@
                         Heroes loses 1 XP per acre lost from invasion, however this loss cannot exceed the XP required to maintain its current level.
                     </p>
                     <p>Basic hero classes can be retired and replaced with another class. Basic hero classes will start with XP equal to half that of its predecessor.</p>
-                    <p>Advanced hero classes cannot be retired, cannot be selected until the 10th day of the round, and unlock additional upgrades.</p>
+                    <p>Advanced hero classes cannot be retired, cannot be selected until the 5th day of the round, and unlock additional upgrades.</p>
                 </div>
                 <div class="col-md-4">
                     <h4>Basic Classes</h4>
@@ -50,7 +50,6 @@
                                     <th>Name</th>
                                     <th>Level Bonus</th>
                                     <th>Requirement</th>
-                                    <th>Starting XP</th>
                                     <th>Upgrades</th>
                                 </tr>
                             </thead>
@@ -62,11 +61,12 @@
                                     <td>{{ $class['name'] }}</td>
                                     <td>{{ ucwords(str_replace('_', ' ', $class['perk_type'])) }}</td>
                                     <td>{{ $heroHelper->getRequirementDisplay($class) }}</td>
-                                    <td>{{ $heroHelper->getStartingExperienceDisplay($class) }}</td>
                                     <td>
                                         @foreach ($heroHelper->getHeroUpgradesByName($class['perks'])->where('active', true) as $upgrade)
                                             {{ $upgrade->name }}
-                                            @if ($upgrade->type !== 'directive')
+                                            @if ($upgrade->type == 'directive')
+                                                (required)
+                                            @else
                                                 (optional)
                                             @endif
                                             <br/>